# MNT Speed up of plot_face_recognition.py example