--- a/examples/applications/plot_face_recognition.py
+++ b/examples/applications/plot_face_recognition.py
@@ -108,7 +108,7 @@ param_grid = {
     "C": [1e3, 5e3, 1e4, 5e4, 1e5],
     "gamma": [0.0001, 0.0005, 0.001, 0.005, 0.01, 0.1],
 }
-clf = GridSearchCV(SVC(kernel="rbf", class_weight="balanced"), param_grid)
+clf = GridSearchCV(SVC(kernel="rbf", class_weight="balanced"), param_grid, n_jobs=-1)
 clf = clf.fit(X_train_pca, y_train)
 print("done in %0.3fs" % (time() - t0))
 print("Best estimator found by grid search:")