--- a/examples/applications/plot_face_recognition.py
+++ b/examples/applications/plot_face_recognition.py
@@ -10,24 +10,8 @@ The dataset used in this example is a preprocessed excerpt of the
 
 .. _LFW: http://vis-www.cs.umass.edu/lfw/
 
-Expected results for the top 5 most represented people in the dataset:
-
-================== ============ ======= ========== =======
-                   precision    recall  f1-score   support
-================== ============ ======= ========== =======
-     Ariel Sharon       0.67      0.92      0.77        13
-     Colin Powell       0.75      0.78      0.76        60
-  Donald Rumsfeld       0.78      0.67      0.72        27
-    George W Bush       0.86      0.86      0.86       146
-Gerhard Schroeder       0.76      0.76      0.76        25
-      Hugo Chavez       0.67      0.67      0.67        15
-       Tony Blair       0.81      0.69      0.75        36
-
-      avg / total       0.80      0.80      0.80       322
-================== ============ ======= ========== =======
-
 """
-
+# %%
 from time import time
 import logging
 import matplotlib.pyplot as plt
@@ -45,9 +29,9 @@ from sklearn.svm import SVC
 logging.basicConfig(level=logging.INFO, format="%(asctime)s %(message)s")
 
 
-# #############################################################################
+# %%
 # Download the data, if not already on disk and load it as numpy arrays
-
+# %%
 lfw_people = fetch_lfw_people(min_faces_per_person=70, resize=0.4)
 
 # introspect the images arrays to find the shapes (for plotting)
@@ -69,18 +53,20 @@ print("n_features: %d" % n_features)
 print("n_classes: %d" % n_classes)
 
 
-# #############################################################################
+# %%
 # Split into a training set and a test set using a stratified k fold
 
 # split into a training and testing set
+# %%
 X_train, X_test, y_train, y_test = train_test_split(
     X, y, test_size=0.25, random_state=42
 )
 
 
-# #############################################################################
+# %%
 # Compute a PCA (eigenfaces) on the face dataset (treated as unlabeled
 # dataset): unsupervised feature extraction / dimensionality reduction
+# %%
 n_components = 150
 
 print(
@@ -99,9 +85,9 @@ X_test_pca = pca.transform(X_test)
 print("done in %0.3fs" % (time() - t0))
 
 
-# #############################################################################
+# %%
 # Train a SVM classification model
-
+# %%
 print("Fitting the classifier to the training set")
 t0 = time()
 param_grid = {
@@ -117,9 +103,9 @@ print("Best estimator found by grid search:")
 print(clf.best_estimator_)
 
 
-# #############################################################################
+# %%
 # Quantitative evaluation of the model quality on the test set
-
+# %%
 print("Predicting people's names on the test set")
 t0 = time()
 y_pred = clf.predict(X_test_pca)
@@ -129,9 +115,9 @@ print(classification_report(y_test, y_pred, target_names=target_names))
 print(confusion_matrix(y_test, y_pred, labels=range(n_classes)))
 
 
-# #############################################################################
+# %%
 # Qualitative evaluation of the predictions using matplotlib
-
+# %%
 
 def plot_gallery(images, titles, h, w, n_row=3, n_col=4):
     """Helper function to plot a gallery of portraits"""