#### Reference Issues/PRs

References: #21598

#### What does this implement/fix? Explain your changes.
`GridSearchCV` now leverages all the cores on the machine by specifying `n_jobs=-1` -- this is the most time consuming step in this example.

This example runs in 6s now which used to take around 20s (3x speedup)

Before:
```
Fitting the classifier to the training set
done in 21.296s
```

After:
```
Fitting the classifier to the training set
done in 6.311s
```