@sveneschlbeck 2021-11-21
@adrinjalali @ogrisel 

Hey Pramod,

thanks for the PR. I appreciate the effort :)
However, we already had several discussions concerning the ``n_jobs`` parameter and decided on not changing it to ``-1`` for concurrency reasons. Let me quote @ogrisel who argued:

_Running with -1 by default is problematic because on machines with a large number of CPUs (e.g. 64 or more), spawning the workers can dominate with concurrent access to the hard disk just to start the python interpreters and import the modules. Furthermore it can also use too much memory and cause crashes.

This is why we would rather use a small number of workers (e.g. 2 instead of -1) when we want to use parallelism in examples or tests in scikit-learn._

Therefore, we might either change the param to ``2`` or start looking for some other ways of speeding up the exec.