@sveneschlbeck 2021-11-21
@pramodatre You could maybe try and alter the ``n_components`` param and see if the results are stable or much worse (maybe go down from 150 to 120 or 100)