@pramodatre 2021-11-21
@sveneschlbeck Thank you for your note on `n_jobs=-1`. So, does that mean we can set `n_jobs=2` then?

Also, I will try to play with `n_components` and report on performance changes.