@sveneschlbeck 2021-11-21
@pramodatre I would wait for a second reviewer but yes, I think that setting ``n_jobs`` to 2 is good. This should improve the exec time. The point is just that we don't want to use ALL ressources available...
Great, waiting for your report :)