@pramodatre 2021-11-21
@sveneschlbeck Sounds good for the second reviewer input for `n_jobs`. In parallel, I will be exploring other options as you suggested. BTW, this is my first PR for an open-source project and I'm thrilled to learn more about the process (I'm already learning!). Thank you for your help!