@sveneschlbeck 2021-11-21
@pramodatre Sounds great! All help is welcome.

Here's a tip for you. If we make improvements or changes to the code, we always back it up by posting some kind of metrics (proof that it did indeed increase speed in this case). This could be done by pasting in a screenshot of your execution showing both the execution time before/after and the output result (plots) before/after. This makes it easy to analyse whether the changes are helpful or not.
In general, we want to decrease the execution time while keeping the results the same, meaning
- Plots should look the same or similar (small differences are no problem)
- Accuracy, Loss Rate or other metrics should no vary to much (0.98 insted of 0.99 would be acceptable but not 0.5 instead of 0.9)
- The main idea or learning goal behind an example should not be falsified (e.g. when the purpose of an example is to show the working of a specific example or parameter than deleting or replacing that very parameter would counteract the learning effect