@pramodatre 2021-11-23
> Let's use `n_jobs=2` instead of `n_jobs=-1`. Please also report the difference in speed you observe when running with `n_jobs=2` vs the default (sequential run).
> 
> Sometimes starting work processes and data communication overhead between processes do no make it possible to observe an actual speed improvement (depending the size of the data and the duration of the parallel tasks).

Sounds good! Just curious of the choice of `n_jobs=2` -- is this decision based on the most common hardware configuration of dual core?