@glemaitre 2021-11-23
The CircleCI machine allocates us only 2 cores if I am not wrong. Here, it should be enough for the task that we want to do indeed. In practice, it would be a bit of a waste to take 40 cores of a big machine for this example :)

In some way, this is more pedagogical to show a user that you should set the value to a meaningful value instead of just expanding to whatever is available.