@pramodatre 2021-11-23
Here are the results of various experiments. In conclusion, `n_components = 150`, `RandomizedSearchCV`, `n_iter=10` gives comparable results to the expected precision/recall/f-score mentioned in the example docstring. The running time for this configuration is around 6s compared to 21s (around 3x speedup).

Please look at the `Averages for all experiments` section in the excel sheet.

[Scikit-learn example plot_face_recognition.py profiling.xlsx](https://github.com/scikit-learn/scikit-learn/files/7587986/Scikit-learn.example.plot_face_recognition.py.profiling.xlsx)