@glemaitre 2021-11-23
@pramodatre Thanks. Could you provide a markdown table with the results instead of an Excel spreadsheet? It would be easier to have a self-contained message in the PR.

I push with the solution is as well a good things because we will run the example and check the speed-up that we get on CircleCI compare to the documentation created for `dev`.