@pramodatre 2021-11-23
> @pramodatre Thanks. Could you provide a markdown table with the results instead of an Excel spreadsheet? It would be easier to have a self-contained message in the PR.
> 
> I push with the solution is as well a good thing because we will run the example and check the speed-up that we get on CircleCI compared to the documentation created for `dev`.

@glemaitre Thanks for the feedback! Yeah, did not realize that excel will hide all the details. Are screenshots of the evaluation tables an option? If not, for sure, I will create markdown. Please let me know.