@pramodatre 2021-11-23
@glemaitre Yeah, good point! I will go ahead and create markdown tables then. Thanks!